error on unknown attributes in derive macros.
This has both false positives and false negatives.

False positives: any unrecognized attribute macro that does not modify the input item in a significant way.
False negatives: attribute proc-macros with the same name as a helper attribute from this crate. We can't detect which it is because we cannot know if the relevant derive macro is actually in use from a separate derive macro.